moved concrete networks into a separate module
