+++ network_diffusion/networks/__init__.py
@@ -1,0 +1,1 @@
+"""This module contains networks which can be used with the library."""
