+++ network_diffusion/networks/data/1_course_edges.csv
