+++ network_diffusion/networks/data/1_ego_edges.csv
