+++ network_diffusion/networks/data/1_nodes.csv
