+++ network_diffusion/networks/data/2_course_edges.csv
