+++ network_diffusion/networks/data/2_ego_edges.csv
