+++ network_diffusion/networks/data/2_nodes.csv
