+++ network_diffusion/networks/data/3_course_edges.csv
