+++ network_diffusion/networks/data/3_ego_edges.csv
