+++ network_diffusion/networks/data/3_nodes.csv
