+++ network_diffusion/networks/l2_course_net.py
